gracc-oneoffs/find-stopped-reporting/requirements.txt
certifi==2021.10.8
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
python-dateutil==2.8.2
six==1.16.0
urllib3==1.26.9
